Refactor database seeders and enhance home view with activity calendar
- Merged SampleDataSeeder functionality into DatabaseSeeder for better organization and to eliminate redundancy.
- Removed SampleDataSeeder file as its functionality is now integrated into DatabaseSeeder.
- Enhanced the home view to include a calendar displaying activities for the current month, improving user experience.
- Added custom styles for the calendar layout and events in the home view.
